requirements.txt
blinker==1.9.0
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.2.0
protobuf==4.21.12
python-dotenv==1.2.1
six==1.17.0
urllib3==2.5.0
Werkzeug==3.0.1
certifi
gunicorn
